Rename MetadataBase -> Metadata::BaseMetadata
This matches the naming convention of Prompters and Renderers.
